Add option to disable font hinting.
Also improve text on non-hidpi screens.

+++ src/tofi.h
@@ -48,7 +48,7 @@ struct tofi {
 	} window;
 
 	/* Options */
-	int8_t anchor;
+	uint32_t anchor;
 	bool hide_cursor;
 	bool use_history;
 };
